Tighten session provisioning boundaries
