+++ crates/loom/src/lib.rs
@@ -37,12 +37,14 @@ pub mod mcp;
 pub mod metadata_assist;
 pub mod monitor;
 pub mod profile;
+pub(crate) mod provision;
 pub mod repo;
 pub mod repo_env;
 pub mod review_delivery;
 pub mod runner;
 pub mod runs;
 pub(crate) mod runtime;
+pub(crate) mod scratch;
 pub mod server;
 pub mod session;
 pub(crate) mod session_layout;
