+++ crates/loom/src/runtime.rs
@@ -1,14 +1,6 @@
-//! Actor-aware boundary for session runtime operations.
-//!
-//! HTTP, GitHub, Slack, watches, and automation runs construct an [`Actor`]
-//! through the narrow constructors here. Session orchestration consumes the
-//! actor's derived attribution instead of accepting client-authored origin or
-//! creator strings.
+//! Shared session runtime policy and environment helpers.
 
-use crate::auth::{Grant, Principal};
-use crate::web::ApiResult;
-use crate::{AppState, Db};
-use weaver_api::{CreateReq, SessionView};
+use crate::Db;
 
 pub(crate) const MISSING_GITHUB_TOKEN_MESSAGE: &str = "No GitHub token configured. Add your personal GitHub token in Settings > Account, or configure a write-only GH_TOKEN on the selected profile.";
 
@@ -80,6 +72,20 @@ pub(crate) async fn layer_launch_environment(
     }
     tracing::debug!(repo = %repo_root_str, profile = profile_name, strict, env_vars = env.len(), "layered launch environment");
     env
+}
+
+pub(crate) async fn launch_environment(
+    db: &Db,
+    repo_root: &std::path::Path,
+    cfg: &weaver_core::repo_config::RepoConfig,
+    profile_name: &str,
+    strict: bool,
+    restricted: bool,
+) -> Vec<(String, String)> {
+    let env = crate::profile::env_pairs(db, profile_name)
+        .await
+        .unwrap_or_default();
+    layer_launch_environment(db, repo_root, cfg, profile_name, env, strict, restricted).await
 }
 
 pub(crate) async fn apply_user_github_token(
@@ -160,153 +166,198 @@ pub(crate) async fn github_token_available(
     Ok(false)
 }
 
-#[derive(Debug, Clone)]
-enum ActorKind {
-    Admin {
-        username: String,
-        delegated: bool,
-    },
-    Producer {
-        origin: &'static str,
-        subject: String,
-    },
-    Automation {
-        origin: String,
-        subject: String,
-        profiles: Vec<String>,
-        run_id: Option<String>,
-        session_id: Option<String>,
-    },
-    Session {
-        username: String,
-        session_id: String,
-        branch_id: String,
-    },
-}
+#[cfg(test)]
+mod tests {
+    use super::*;
 
-#[derive(Debug, Clone)]
-pub(crate) struct Actor(ActorKind);
+    async fn seed_user(db: &Db, username: &str) {
+        sqlx::query("INSERT INTO users (username) VALUES (?)")
+            .bind(username)
+            .execute(db)
+            .await
+            .unwrap();
+    }
 
-impl Actor {
-    pub(crate) fn from_principal(principal: &Principal, delegated: bool) -> Self {
-        match &principal.grant {
-            Grant::Admin => Self(ActorKind::Admin {
-                username: principal.username.clone(),
-                delegated,
-            }),
-            Grant::Automation { subject, profiles } => Self(ActorKind::Automation {
-                origin: "automation".to_string(),
-                subject: subject.clone(),
-                profiles: profiles.clone(),
-                run_id: None,
-                session_id: None,
-            }),
-            Grant::Session {
-                session_id,
-                branch_id,
-            } => Self(ActorKind::Session {
-                username: principal.username.clone(),
-                session_id: session_id.clone(),
-                branch_id: branch_id.clone(),
-            }),
+    struct EnvVarGuard {
+        name: &'static str,
+        value: Option<std::ffi::OsString>,
+    }
+
+    impl EnvVarGuard {
+        fn unset(name: &'static str) -> Self {
+            let value = std::env::var_os(name);
+            std::env::remove_var(name);
+            Self { name, value }
+        }
+
+        fn set(name: &'static str, new_value: &str) -> Self {
+            let value = std::env::var_os(name);
+            std::env::set_var(name, new_value);
+            Self { name, value }
         }
     }
 
-    pub(crate) fn producer(origin: &'static str, subject: impl Into<String>) -> Self {
-        debug_assert!(matches!(
-            origin,
-            "github" | "slack" | "watch" | "monitor" | "startup"
-        ));
-        Self(ActorKind::Producer {
-            origin,
-            subject: subject.into(),
-        })
-    }
-
-    pub(crate) fn automation(
-        origin: impl Into<String>,
-        subject: impl Into<String>,
-        profiles: Vec<String>,
-        run_id: impl Into<String>,
-        session_id: impl Into<String>,
-    ) -> Self {
-        Self(ActorKind::Automation {
-            origin: origin.into(),
-            subject: subject.into(),
-            profiles,
-            run_id: Some(run_id.into()),
-            session_id: Some(session_id.into()),
-        })
-    }
-
-    pub(crate) fn origin(&self) -> &str {
-        match &self.0 {
-            ActorKind::Admin {
-                delegated: true, ..
-            }
-            | ActorKind::Session { .. } => "agent",
-            ActorKind::Admin { .. } => "user",
-            ActorKind::Producer { origin, .. } => origin,
-            ActorKind::Automation { origin, .. } => origin,
-        }
-    }
-
-    pub(crate) fn display_creator(&self) -> Option<String> {
-        match &self.0 {
-            ActorKind::Admin { username, .. } | ActorKind::Session { username, .. } => {
-                Some(username.clone())
-            }
-            ActorKind::Producer { subject, .. } | ActorKind::Automation { subject, .. } => {
-                Some(subject.clone())
+    impl Drop for EnvVarGuard {
+        fn drop(&mut self) {
+            match &self.value {
+                Some(value) => std::env::set_var(self.name, value),
+                None => std::env::remove_var(self.name),
             }
         }
     }
 
-    pub(crate) fn bound_parent_branch(&self) -> Option<&str> {
-        match &self.0 {
-            ActorKind::Session { branch_id, .. } => Some(branch_id),
-            _ => None,
-        }
+    #[tokio::test]
+    async fn launch_environment_respects_restricted_and_strict_profiles() {
+        let db = crate::db::connect_in_memory().await.unwrap();
+        let dir = tempfile::tempdir().unwrap();
+        let repo = dir.path();
+        crate::profile::env_set(&db, "github_comment", "GH_TOKEN", "profile-token")
+            .await
+            .unwrap();
+        crate::repo_env::set(&db, &repo.display().to_string(), "REPO_SECRET", "leak")
+            .await
+            .unwrap();
+        let mut restricted_cfg = weaver_core::repo_config::RepoConfig::default();
+        restricted_cfg
+            .env
+            .insert("COMMITTED_SECRET".to_string(), "leak".to_string());
+        assert_eq!(
+            launch_environment(&db, repo, &restricted_cfg, "github_comment", true, true).await,
+            vec![("GH_TOKEN".to_string(), "profile-token".to_string())]
+        );
+
+        crate::profile::env_set(&db, "default", "SHARED_TOKEN", "profile")
+            .await
+            .unwrap();
+        crate::repo_env::set(
+            &db,
+            &repo.display().to_string(),
+            "SHARED_TOKEN",
+            "repository",
+        )
+        .await
+        .unwrap();
+        let mut strict_cfg = weaver_core::repo_config::RepoConfig::default();
+        strict_cfg
+            .env
+            .insert("SHARED_TOKEN".to_string(), "committed".to_string());
+        let strict = launch_environment(&db, repo, &strict_cfg, "default", true, false).await;
+        assert_eq!(
+            strict
+                .iter()
+                .find(|(name, _)| name == "SHARED_TOKEN")
+                .map(|(_, value)| value.as_str()),
+            Some("profile")
+        );
+        assert!(!strict.iter().any(|(name, _)| name == "CARGO_TARGET_DIR"));
     }
 
-    pub(crate) fn creator_identity(&self) -> (&'static str, String) {
-        match &self.0 {
-            ActorKind::Admin { username, .. } => ("user", username.clone()),
-            ActorKind::Producer { subject, .. } => ("system", subject.clone()),
-            ActorKind::Automation { subject, .. } => ("automation", subject.clone()),
-            ActorKind::Session { session_id, .. } => ("session", session_id.clone()),
-        }
+    #[tokio::test]
+    async fn user_github_token_overlay_preserves_precedence() {
+        let db = crate::db::connect_in_memory().await.unwrap();
+        seed_user(&db, "alice").await;
+        seed_user(&db, "bob").await;
+        crate::user_token::set(&db, "alice", "ghp_alice")
+            .await
+            .unwrap();
+
+        let mut fresh = vec![("FOO".to_string(), "bar".to_string())];
+        apply_user_github_token(&db, &mut fresh, Some("alice")).await;
+        assert!(fresh
+            .iter()
+            .any(|(name, value)| name == "GH_TOKEN" && value == "ghp_alice"));
+
+        let mut layered = vec![("GH_TOKEN".to_string(), "shared".to_string())];
+        apply_user_github_token(&db, &mut layered, Some("alice")).await;
+        let github_tokens = layered
+            .iter()
+            .filter(|(name, _)| name == "GH_TOKEN")
+            .collect::<Vec<_>>();
+        assert_eq!(github_tokens.len(), 1);
+        assert_eq!(github_tokens[0].1, "ghp_alice");
+
+        let mut fallback = vec![("GH_TOKEN".to_string(), "shared".to_string())];
+        apply_user_github_token(&db, &mut fallback, Some("bob")).await;
+        assert_eq!(fallback[0].1, "shared");
+        let mut producer = Vec::new();
+        apply_user_github_token(&db, &mut producer, None).await;
+        assert!(producer.is_empty());
     }
 
-    pub(crate) fn allowed_profiles(&self) -> Option<&[String]> {
-        match &self.0 {
-            ActorKind::Automation { profiles, .. } => Some(profiles),
-            _ => None,
-        }
-    }
+    #[serial_test::serial]
+    #[tokio::test]
+    async fn github_token_preflight_covers_builtin_custom_and_app_paths() {
+        let ambient = EnvVarGuard::unset("GH_TOKEN");
+        let db = crate::db::connect_in_memory().await.unwrap();
+        seed_user(&db, "alice").await;
 
-    pub(crate) fn automation_run_id(&self) -> Option<&str> {
-        match &self.0 {
-            ActorKind::Automation { run_id, .. } => run_id.as_deref(),
-            _ => None,
-        }
-    }
+        assert!(
+            !github_token_available(&db, &[], Some("alice"), "codex", None)
+                .await
+                .unwrap()
+        );
+        assert!(
+            github_token_available(&db, &[], Some("alice"), "my-custom-agent", None)
+                .await
+                .unwrap()
+        );
+        assert!(
+            github_token_available(&db, &[], Some("github-webhook (octo)"), "codex", None)
+                .await
+                .unwrap()
+        );
 
-    pub(crate) fn reserved_session_id(&self) -> Option<&str> {
-        match &self.0 {
-            ActorKind::Automation { session_id, .. } => session_id.as_deref(),
-            _ => None,
-        }
-    }
-}
+        crate::user_token::set(&db, "alice", "ghp_alice")
+            .await
+            .unwrap();
+        assert!(
+            github_token_available(&db, &[], Some("alice"), "claude", None)
+                .await
+                .unwrap()
+        );
+        crate::user_token::remove(&db, "alice").await.unwrap();
+        assert!(github_token_available(
+            &db,
+            &[("GH_TOKEN".to_string(), "ghp_shared".to_string())],
+            Some("alice"),
+            "codex",
+            None,
+        )
+        .await
+        .unwrap());
 
-/// The single actor-taking entrypoint for all session producers. HTTP, Slack,
-/// GitHub, and automation runs cannot bypass attribution/grant derivation by
-/// calling the web provisioning implementation directly.
-pub(crate) async fn create_session(
-    state: AppState,
-    request: CreateReq,
-    actor: Actor,
-) -> ApiResult<SessionView> {
-    crate::web::sessions::provision_session(state, request, actor).await
+        weaver_core::config::apply(
+            &db,
+            &[
+                (
+                    crate::github_app::APP_ID_KEY.to_string(),
+                    Some("123456".to_string()),
+                ),
+                (
+                    crate::github_app::APP_PRIVATE_KEY_KEY.to_string(),
+                    Some("configured-for-preflight".to_string()),
+                ),
+            ],
+        )
+        .await
+        .unwrap();
+        let app = crate::github_app::GithubApp::new(db.clone());
+        assert!(
+            github_token_available(&db, &[], Some("alice"), "claude", Some(&app))
+                .await
+                .unwrap()
+        );
+
+        drop(ambient);
+        let _ambient = EnvVarGuard::set("GH_TOKEN", "ghp_ambient");
+        assert!(!github_token_available(
+            &db,
+            &[("GH_TOKEN".to_string(), " ".to_string())],
+            Some("alice"),
+            "codex",
+            None,
+        )
+        .await
+        .unwrap());
+    }
 }
