+++ crates/loom/src/web/sessions.rs
@@ -21,7 +21,7 @@ use crate::auth::Principal;
 use crate::db::Db;
 use crate::events::Event;
 use crate::session::{self as session_mod, NewSession, Session};
-use crate::{agent, backend, config, custom_agents, db, events, git, github, repo, setup};
+use crate::{agent, backend, config, custom_agents, db, events, git, github, repo};
 use weaver_api::{
     CreateReq, EnsureResumptionCueReq, HandoffReq, HistoryPageView, LaunchOverrides,
     LaunchSelection, PatchSessionReq, ResumptionCueView, SearchSessionsOptions, SendReq,
@@ -33,29 +33,13 @@ use weaver_core::branch::{Branch, TitleProvenance, TitleUpdate};
 use weaver_core::tags;
 use weaver_core::watch::{self as watch_store, Watch};
 
-use super::scratch::{prepare_initial_scratch, scratch_note, write_prepared_initial_scratch};
 use super::{author_or_manual, require_branch, require_session, session_view};
 use super::{ApiResult, AppError, AppState};
-use crate::runtime::{
-    apply_user_github_token, layer_launch_environment, repo_cfg_or_default, set_env,
-};
-
-async fn insert_session_row(
-    st: &AppState,
-    session: &NewSession,
-    policy: &session_mod::SessionLaunchPolicy,
-) -> Result<Session, AppError> {
-    let (session, layout_revision) =
-        session_mod::insert_with_layout_revision(&st.db, session, policy).await?;
-    if let Some(revision) = layout_revision {
-        super::session_layout::publish_invalidation(st, revision).await;
-    }
-    Ok(session)
-}
+use crate::runtime::{layer_launch_environment, repo_cfg_or_default, set_env};
 
 async fn delete_session_row(st: &AppState, session_id: &str) -> Result<(), AppError> {
     if let Some(revision) = session_mod::delete(&st.db, session_id).await? {
-        super::session_layout::publish_invalidation(st, revision).await;
+        crate::session_layout::publish_invalidation(&st.db, &st.bus, revision).await;
     }
     Ok(())
 }
@@ -351,7 +335,8 @@ pub(super) async fn create_session(
 ) -> ApiResult<Json<SessionView>> {
     // The authenticated managed-repository convenience below is itself a
     // durable side effect. Reject the complete untrusted Scratch batch first.
-    let _ = prepare_initial_scratch(&req.scratch)?;
+    let _ = crate::scratch::prepare_initial_scratch(&req.scratch)
+        .map_err(super::scratch::map_scratch_error)?;
     // Naming a managed repo here registers it: a signed-in principal asking to
     // launch into `owner/name` is the grant, so a repo loom has never seen just
     // works (it is cloned on the way through runtime provisioning). The `repos`
@@ -374,7 +359,7 @@ pub(super) async fn create_session(
         .as_deref()
         .map(str::trim)
         .is_some_and(|value| !value.is_empty());
-    let actor = crate::runtime::Actor::from_principal(&principal, delegated);
+    let actor = crate::provision::Actor::from_principal(&principal, delegated);
     match &principal.grant {
         crate::auth::Grant::Session { .. } => {
             req.parent_branch = actor.bound_parent_branch().map(str::to_string);
@@ -387,7 +372,12 @@ pub(super) async fn create_session(
         }
         crate::auth::Grant::Admin => {}
     }
-    Ok(Json(crate::runtime::create_session(st, req, actor).await?))
+    let created = crate::provision::create(st.clone(), req, actor)
+        .await
+        .map_err(super::provision_error)?;
+    Ok(Json(
+        session_view(&st.db, &created.session, &created.branch).await?,
+    ))
 }
 
 /// Add a managed-repo reference to the registry if it isn't there yet — the same
@@ -409,27 +399,6 @@ async fn ensure_repo_registered(db: &Db, input: &str) -> ApiResult<()> {
     Ok(())
 }
 
-/// Build the environment exported into a session's agent terminal, starting
-/// with the selected profile and then layering the per-repo [`repo_env`] and
-/// committed `.weaver/config.toml` `[env]`. A strict profile keeps ownership of
-/// its declared names; a restricted profile receives only its own environment.
-/// Loom fills in its repo-local defaults last only when no layer supplied the
-/// name. Best-effort: a database error in a layer degrades to the layers that
-/// did resolve. `cfg` is the already-loaded repo config.
-async fn launch_env_for_profile(
-    db: &Db,
-    repo_root: &std::path::Path,
-    cfg: &weaver_core::repo_config::RepoConfig,
-    profile_name: &str,
-    strict: bool,
-    restricted: bool,
-) -> Vec<(String, String)> {
-    let env = crate::profile::env_pairs(db, profile_name)
-        .await
-        .unwrap_or_default();
-    layer_launch_environment(db, repo_root, cfg, profile_name, env, strict, restricted).await
-}
-
 /// Build the explicit ambient baseline used when Tapestry clears inheritance.
 /// Profile/repo values win over baseline and allowlisted ambient values; loom's
 /// own session variables are injected later by `agent::session_env`.
@@ -439,7 +408,7 @@ async fn resume_environment(
     repo_root: &std::path::Path,
     cfg: &weaver_core::repo_config::RepoConfig,
 ) -> Vec<(String, String)> {
-    let env = launch_env_for_profile(
+    let env = crate::runtime::launch_environment(
         db,
         repo_root,
         cfg,
@@ -473,189 +442,6 @@ async fn rotate_session_token(
     Ok(())
 }
 
-async fn ensure_github_token_available(
-    db: &Db,
-    env: &[(String, String)],
-    created_by: Option<&str>,
-    runtime: &str,
-    restricted_github_app: Option<&crate::github_app::GithubApp>,
-) -> ApiResult<()> {
-    if crate::runtime::github_token_available(db, env, created_by, runtime, restricted_github_app)
-        .await?
-    {
-        Ok(())
-    } else {
-        Err(AppError::new(
-            StatusCode::PRECONDITION_REQUIRED,
-            crate::runtime::MISSING_GITHUB_TOKEN_MESSAGE,
-        ))
-    }
-}
-
-async fn fetch_launch_issue(
-    st: &AppState,
-    repo_root: &std::path::Path,
-    managed_repo: Option<&crate::repo::RepoSlug>,
-    number: i64,
-) -> anyhow::Result<github::Issue> {
-    if let (Some(app), Some(repo)) = (st.trigger.app(), managed_repo) {
-        if app.is_configured().await {
-            return app.issue(&repo.owner, &repo.name, number).await;
-        }
-    }
-    github::fetch_issue(repo_root, number).await
-}
-
-/// The configured wall-clock budget for a repo setup run.
-async fn setup_timeout(db: &Db) -> std::time::Duration {
-    let secs = config::get(db, "setup.timeout_secs")
-        .await
-        .and_then(|v| v.trim().parse::<u64>().ok())
-        .unwrap_or(config::DEFAULT_SETUP_TIMEOUT_SECS as u64)
-        .max(1);
-    std::time::Duration::from_secs(secs)
-}
-
-fn repo_setup_for_profile(
-    cfg: &weaver_core::repo_config::RepoConfig,
-    restricted: bool,
-) -> Option<String> {
-    (!restricted).then(|| cfg.setup.script()).flatten()
-}
-
-/// Run a registered repo's `[setup]` script in the worktree before the agent
-/// starts, recording its lifecycle as `setup` events (so the session view shows
-/// it) and capturing full output to `setup.log` in the run dir. The caller has
-/// already confirmed the repo is allowlisted. Returns the outcome; the caller
-/// decides whether to launch the agent or leave the session in an error state.
-async fn run_repo_setup(
-    st: &AppState,
-    branch_id: &str,
-    work_dir: &std::path::Path,
-    run_dir: &std::path::Path,
-    script: &str,
-    env: &[(String, String)],
-) -> setup::SetupOutcome {
-    let timeout = setup_timeout(&st.db).await;
-    tracing::info!(branch = branch_id, work_dir = %work_dir.display(), timeout_secs = timeout.as_secs(), "running repo [setup] script");
-    events::record(
-        &st.db,
-        &st.bus,
-        branch_id,
-        "setup",
-        json!({ "phase": "started", "timeout_secs": timeout.as_secs() }),
-    )
-    .await
-    .ok();
-
-    let log_path = run_dir.join("setup.log");
-    let outcome = setup::run(work_dir, script, env, timeout, Some(&log_path))
-        .await
-        .unwrap_or_else(|e| setup::SetupOutcome {
-            success: false,
-            timed_out: false,
-            exit_code: None,
-            output: format!("failed to start setup: {e}"),
-            duration: std::time::Duration::ZERO,
-        });
-
-    // The full output lives in setup.log; the event carries a bounded tail so the
-    // timeline stays light.
-    let tail = tail_chars(&outcome.output, 4000);
-    events::record(
-        &st.db,
-        &st.bus,
-        branch_id,
-        "setup",
-        json!({
-            "phase": "finished",
-            "success": outcome.success,
-            "timed_out": outcome.timed_out,
-            "exit_code": outcome.exit_code,
-            "duration_ms": outcome.duration.as_millis() as u64,
-            "summary": outcome.summary(),
-            "output": tail,
-        }),
-    )
-    .await
-    .ok();
-    if outcome.success {
-        tracing::info!(branch = branch_id, "repo setup succeeded");
-    } else {
-        tracing::warn!(branch = branch_id, summary = %outcome.summary(), "repo setup failed");
-    }
-    outcome
-}
-
-/// The last `max` chars of `s` (whole string when shorter), prefixed with an
-/// elision marker when truncated. Keeps a setup-output event payload bounded.
-fn tail_chars(s: &str, max: usize) -> String {
-    if s.chars().count() <= max {
-        return s.to_string();
-    }
-    let tail: String = s
-        .chars()
-        .rev()
-        .take(max)
-        .collect::<Vec<_>>()
-        .into_iter()
-        .rev()
-        .collect();
-    format!("…(truncated)\n{tail}")
-}
-
-fn legacy_launch_selection(req: &CreateReq) -> LaunchSelection {
-    let nonempty = |value: &Option<String>| {
-        value
-            .as_ref()
-            .map(|value| value.trim().to_string())
-            .filter(|value| !value.is_empty())
-    };
-    let selected = |value: &Option<String>| value.as_ref().map(|value| value.trim().to_string());
-    LaunchSelection {
-        profile: nonempty(&req.profile)
-            .unwrap_or_else(|| crate::profile::DEFAULT_PROFILE.to_string()),
-        overrides: LaunchOverrides {
-            agent: nonempty(&req.agent),
-            // Empty is an explicit "agent default" selector for these legacy
-            // fields, distinct from omission (which inherits the profile).
-            model: selected(&req.model),
-            effort: selected(&req.effort),
-            protocol: selected(&req.protocol),
-            mode: nonempty(&req.mode),
-            class: nonempty(&req.class),
-        },
-    }
-}
-
-fn create_selection(req: &CreateReq) -> ApiResult<LaunchSelection> {
-    let Some(selection) = &req.selection else {
-        return Ok(legacy_launch_selection(req));
-    };
-    let flattened = [
-        &req.profile,
-        &req.agent,
-        &req.model,
-        &req.effort,
-        &req.protocol,
-        &req.mode,
-        &req.class,
-    ]
-    .into_iter()
-    .any(Option::is_some);
-    if flattened {
-        return Err(AppError::bad_request(
-            "canonical `selection` cannot be combined with flattened launch selectors",
-        ));
-    }
-    if req.expected_profile_revision.is_none() || req.expected_resolver_revision.is_none() {
-        return Err(AppError::bad_request(
-            "canonical `selection` requires expected_profile_revision and expected_resolver_revision from a resolve preview",
-        ));
-    }
-    Ok(selection.clone())
-}
-
 fn map_handoff_error(error: crate::handoff::HandoffError) -> AppError {
     use crate::handoff::HandoffError::*;
     let (status, message, preview) = match error {
@@ -687,1048 +473,6 @@ pub(super) async fn resolve_session_handoff(
         .await
         .map_err(map_handoff_error)?;
     Ok(Json(view))
-}
-
-/// The session-creation core shared by every producer. The actor supplies trusted
-/// attribution, origin, ancestry, and profile bounds; request fields cannot
-/// impersonate those properties. Returns the view directly so each caller can
-/// shape its own response.
-pub(crate) async fn provision_session(
-    st: AppState,
-    req: CreateReq,
-    actor: crate::runtime::Actor,
-) -> ApiResult<SessionView> {
-    let created_by = actor.display_creator();
-    let origin = actor.origin();
-    tracing::info!(
-        repo = ?req.repo,
-        agent = ?req.agent,
-        created_by = ?created_by,
-        origin,
-        "provision_session: starting session creation"
-    );
-    // Attachment input is untrusted launch input, not a provisioning step.
-    // Decode and validate the entire batch before touching a repository,
-    // worktree, branch, tracking issue, claim, or session row.
-    let prepared_scratch = prepare_initial_scratch(&req.scratch)?;
-    let selection = create_selection(&req)?;
-    let selected_profile_name = match selection.profile.trim() {
-        "" => crate::profile::DEFAULT_PROFILE,
-        name => name,
-    }
-    .to_string();
-    // This is both the capacity gate and the template-lifetime gate. Profile
-    // edits, retirement, recreation, clone, and environment mutation use the
-    // same permit; retaining it through session insertion closes the
-    // resolve/provision/delete gap even for unlimited profiles.
-    let _profile_permit = st.launch_gate.acquire_profile(&selected_profile_name).await;
-    // Custom-agent and custom-MCP mutations share this boundary. Keep the
-    // exact registry generation accepted below through command construction
-    // and runtime startup.
-    let _resolver_permit = st.launch_gate.acquire_resolver().await;
-    let options = crate::launch::ResolveOptions {
-        default_class: matches!(origin, "watch" | "actions" | "ops")
-            .then(|| "automation".to_string()),
-        ..Default::default()
-    };
-    let resolved = match super::launches::resolve_launch(&st, &selection, &options).await {
-        Ok(resolved) => resolved,
-        Err(error) if req.expected_resolver_revision.is_some() => {
-            return Err(AppError::conflict(format!(
-                "launch settings can no longer be resolved after preview: {}",
-                error.message()
-            )));
-        }
-        Err(error) => return Err(error),
-    };
-    let stale_profile = req
-        .expected_profile_revision
-        .is_some_and(|expected| expected != resolved.view.profile_revision);
-    let stale_resolver = req
-        .expected_resolver_revision
-        .as_deref()
-        .is_some_and(|expected| expected != resolved.view.resolver_revision);
-    if stale_profile || stale_resolver {
-        return Err(AppError::conflict(
-            "launch settings changed since preview; review the fresh resolution",
-        )
-        .with_fields(json!({ "preview": resolved.view })));
-    }
-    if !resolved.view.valid {
-        let message = resolved
-            .view
-            .errors
-            .first()
-            .cloned()
-            .unwrap_or_else(|| "launch settings are not valid".to_string());
-        let status = if resolved.view.capacity.allowed {
-            StatusCode::BAD_REQUEST
-        } else {
-            StatusCode::CONFLICT
-        };
-        return Err(AppError::new(status, message).with_fields(json!({ "preview": resolved.view })));
-    }
-    // Resolve write-only values once, then confirm the template revision still
-    // matches. Environment mutations advance that revision transactionally, so
-    // provisioning below uses this concrete snapshot instead of silently
-    // reading a value changed after preview.
-    let profile_environment = crate::profile::env_pairs(&st.db, &resolved.profile.name)
-        .await
-        .map_err(|error| AppError::bad_request(error.to_string()))?;
-    let current_profile = crate::profile::get(&st.db, &resolved.profile.name)
-        .await?
-        .ok_or_else(|| AppError::bad_request("selected profile was removed after preview"))?;
-    if current_profile.revision != resolved.view.profile_revision
-        || current_profile.lifetime != resolved.view.profile_lifetime
-    {
-        let fresh = super::launches::resolve_launch(&st, &selection, &options).await?;
-        return Err(AppError::conflict(
-            "launch profile changed while resolving its environment; review the fresh resolution",
-        )
-        .with_fields(json!({ "preview": fresh.view })));
-    }
-    let profile_name = resolved.profile.name.clone();
-    let custom_agent = resolved.custom_agent.clone();
-    let launch_profile = resolved.profile;
-    let agent = resolved.view.agent.clone();
-    let model = resolved.view.model.clone();
-    let effort = resolved.view.effort.clone();
-    let protocol = resolved.view.protocol.clone();
-    let mode = resolved.view.mode.clone();
-    let class = resolved.view.class.clone();
-    let launch_snapshot =
-        crate::launch::serialize_snapshot(&resolved.view, resolved.custom_agent.as_ref())
-            .map_err(|error| AppError::bad_request(error.to_string()))?;
-    if let Some(allowed) = actor.allowed_profiles() {
-        if !allowed.iter().any(|name| name == &profile_name) {
-            return Err(AppError::new(
-                StatusCode::FORBIDDEN,
-                format!("automation grant does not allow profile '{profile_name}'"),
-            ));
-        }
-        if !launch_profile.is_automation_safe() {
-            return Err(AppError::bad_request(format!(
-                "automation profile '{profile_name}' must be automation-class, strict, and env-cleared"
-            )));
-        }
-    }
-    let managed_repo = req.repo.as_deref().map(str::trim).filter(|s| !s.is_empty());
-    // Resolve the stable repository identity without mutating it, then hold its
-    // launch permit through clone/fetch, worktree setup, and agent startup. A
-    // different repository gets a different permit and remains independent.
-    let repo_key = match managed_repo {
-        Some(input) => repo::registered_path(&st.db, input)
-            .await
-            .map_err(|e| match e {
-                repo::ResolveError::BadRequest(m) => AppError::bad_request(m),
-                repo::ResolveError::Clone(m) => AppError::new(StatusCode::BAD_GATEWAY, m),
-            })?,
-        None => {
-            let cwd = PathBuf::from(&req.cwd);
-            git::repo_root(&cwd)
-                .await
-                .map_err(|e| AppError::bad_request(e.to_string()))?
-        }
-    };
-    let repo_key = repo_key.canonicalize().unwrap_or(repo_key);
-    tracing::debug!(repo = %repo_key.display(), "waiting for repository launch gate");
-    let launch_permit = st.launch_gate.acquire(&repo_key).await;
-    tracing::debug!(repo = %repo_key.display(), "acquired repository launch gate");
-
-    // Recheck capacity while holding the profile-wide admission gate. The
-    // permit remains live through session insertion, so launches against
-    // different repositories cannot over-admit the same profile.
-    if launch_profile.max_concurrent > 0
-        && crate::profile::active_count(&st.db, &profile_name).await?
-            >= launch_profile.max_concurrent
-    {
-        let fresh = super::launches::resolve_launch(&st, &selection, &options).await?;
-        return Err(AppError::conflict(format!(
-            "profile '{profile_name}' has reached its max_concurrent limit ({})",
-            launch_profile.max_concurrent
-        ))
-        .with_fields(json!({ "preview": fresh.view })));
-    }
-
-    // Now acquire the managed clone (inside the gate), or reuse the local root
-    // resolved above. The traversal / allowlist boundary lives in `repo`.
-    let repo_root = match managed_repo {
-        Some(input) => repo::resolve_clone(&st.db, input, st.trigger.app())
-            .await
-            .map_err(|e| match e {
-                repo::ResolveError::BadRequest(m) => AppError::bad_request(m),
-                repo::ResolveError::Clone(m) => AppError::new(StatusCode::BAD_GATEWAY, m),
-            })?,
-        None => repo_key,
-    };
-    // Canonicalize so repo identity matches the `weaver` CLI's resolver — issues
-    // are keyed on this path and the two binaries must agree on it.
-    let repo_root = repo_root.canonicalize().unwrap_or(repo_root);
-    tracing::debug!(repo_root = %repo_root.display(), "resolved repo root");
-
-    // The repo's committed `.weaver/config.toml`, read from its primary checkout.
-    // It supplies agent/model/effort defaults (below an explicit request, above
-    // the operator's global default), the `[env]` layer exported into the
-    // terminal, and the `[setup]` bootstrap run for allowlisted repos. A malformed
-    // file is a hard error *only* for an allowlisted repo (whose setup would run),
-    // so the breakage is visible at create time; for any other repo it would have
-    // supplied mere defaults, so we log and proceed with an empty config.
-    let repo_cfg = match weaver_core::repo_config::load(&repo_root) {
-        Ok(cfg) => cfg,
-        Err(e) => {
-            if repo::is_allowlisted(&st.db, &repo_root)
-                .await
-                .unwrap_or(false)
-            {
-                return Err(AppError::bad_request(format!(
-                    "repo .weaver/config.toml is invalid: {e}"
-                )));
-            }
-            tracing::warn!(repo = %repo_root.display(), error = %e,
-                "ignoring malformed .weaver/config.toml");
-            weaver_core::repo_config::RepoConfig::default()
-        }
-    };
-    tracing::debug!(repo_root = %repo_root.display(), "loaded repo config");
-
-    // A GitHub App token is repository-scoped. A managed slug gives both the
-    // preflight and issue seeding an exact installation target; a local path
-    // must keep using an explicitly supplied session credential.
-    let managed_slug = req
-        .repo
-        .as_deref()
-        .and_then(|repo| crate::repo::parse_slug(repo).ok());
-    let runtime = agent.clone();
-    tracing::debug!(agent = %agent, runtime = %runtime, "resolved agent runtime");
-    // The resolved launch environment: selected profile < per-repo repo_env <
-    // the repo file's [env]. It is needed before provisioning so a real agent
-    // launch can stop cleanly when neither the user nor an environment layer
-    // provides GH_TOKEN.
-    let mut extra_env = layer_launch_environment(
-        &st.db,
-        &repo_root,
-        &repo_cfg,
-        &profile_name,
-        profile_environment,
-        launch_profile.strict,
-        launch_profile.restricted,
-    )
-    .await;
-    if launch_profile.env_clear {
-        let allowlist = launch_profile
-            .ambient_names()
-            .map_err(|e| AppError::bad_request(e.to_string()))?;
-        extra_env = crate::profile::cleared_environment(extra_env, &allowlist);
-    }
-    // Select the launching user's GitHub credential by overlaying it as
-    // GH_TOKEN (design §6.3, "Level B"). See `apply_user_github_token` for the
-    // precedence rules. This happens before preflight. Ordinary sessions export
-    // it; a restricted ACP launch removes it from the adapter environment and
-    // its server-side GitHub tool independently resolves an App or profile
-    // credential.
-    apply_user_github_token(&st.db, &mut extra_env, created_by.as_deref()).await;
-
-    tracing::debug!(model = %model, effort = %effort, protocol = %protocol, "resolved and validated model/effort/protocol");
-    let restricted_github_app = if launch_profile.restricted && managed_slug.is_some() {
-        st.trigger.app()
-    } else {
-        None
-    };
-    ensure_github_token_available(
-        &st.db,
-        &extra_env,
-        created_by.as_deref(),
-        &runtime,
-        restricted_github_app,
-    )
-    .await?;
-    tracing::debug!(runtime = %runtime, "github token availability check passed");
-
-    // Build title/goal/description; an optional GitHub issue seeds all three.
-    let mut goal = req.goal.unwrap_or_default().trim().to_string();
-    let mut title = req
-        .title
-        .as_deref()
-        .and_then(branch_mod::sanitize_user_title);
-    let title_was_explicit = title.is_some();
-    let mut description = String::new();
-    let mut github_repo = None;
-    let mut github_issue: Option<i64> = None;
-    if let Some(number) = req.issue {
-        tracing::info!(issue = number, repo = %repo_root.display(), "fetching github issue to seed session");
-        let issue = fetch_launch_issue(&st, &repo_root, managed_slug.as_ref(), number)
-            .await
-            .map_err(|e| AppError::bad_request(format!("issue #{number}: {e}")))?;
-        if title.is_none() {
-            title = Some(issue.title.clone());
-        }
-        if goal.is_empty() {
-            goal = if issue.body.trim().is_empty() {
-                issue.title.clone()
-            } else {
-                format!("{}\n\n{}", issue.title, issue.body)
-            };
-        }
-        description = issue.body.clone();
-        github_issue = Some(number);
-        github_repo = match managed_slug.as_ref() {
-            Some(repo) => Some(repo.slug()),
-            None => github::repo_slug(&repo_root).await.ok(),
-        };
-        tracing::debug!(issue = number, github_repo = ?github_repo, "seeded session fields from github issue");
-    } else if let Some(number) = req.github_issue {
-        // The caller already holds the thread (the `@loom` trigger): record the
-        // GitHub link on the tracking issue without the fetch-and-seed above.
-        github_issue = Some(number);
-        github_repo = managed_slug.as_ref().map(|repo| repo.slug());
-    }
-
-    // Claiming an existing weaver issue seeds the same three fields from it.
-    let repo_root_str = repo_root.display().to_string();
-    let mut claimed_issue_id: Option<i64> = None;
-    if let Some(issue_id) = req.claim_issue {
-        tracing::debug!(issue_id, "claiming existing weaver issue for new session");
-        let issue = weaver_core::issue::get(&st.db, issue_id)
-            .await?
-            .ok_or_else(|| AppError::not_found("issue"))?;
-        if issue.repo_root != repo_root_str {
-            return Err(AppError::bad_request(format!(
-                "issue #{issue_id} belongs to a different repo"
-            )));
-        }
-        if title.is_none() {
-            title = Some(issue.title.clone());
-        }
-        if goal.is_empty() {
-            goal = if issue.body.trim().is_empty() {
-                issue.title.clone()
-            } else {
-                format!("{}\n\n{}", issue.title, issue.body)
-            };
-        }
-        if description.is_empty() {
-            description = issue.body.clone();
-        }
-        claimed_issue_id = Some(issue_id);
-    }
-    let issue_owned_title =
-        req.issue.is_some() || req.github_issue.is_some() || req.claim_issue.is_some();
-    let title_provenance = if title_was_explicit {
-        if issue_owned_title && origin == "github" {
-            TitleProvenance::Issue
-        } else {
-            TitleProvenance::User
-        }
-    } else if issue_owned_title {
-        TitleProvenance::Issue
-    } else {
-        TitleProvenance::Derived
-    };
-    let mut title = title
-        .and_then(|title| branch_mod::sanitize_user_title(&title))
-        .unwrap_or_else(|| branch_mod::derive_title(&goal));
-
-    let existing = req
-        .existing_branch
-        .as_deref()
-        .map(str::trim)
-        .filter(|b| !b.is_empty());
-    if existing.is_some()
-        && req
-            .name
-            .as_deref()
-            .map(str::trim)
-            .is_some_and(|n| !n.is_empty())
-    {
-        return Err(AppError::bad_request(
-            "`name` and `existing_branch` are mutually exclusive",
-        ));
-    }
-
-    // Unless the caller pins a base, fork from a freshly-fetched `origin/<default
-    // branch>` so new work starts from the latest mainline, not the launching
-    // checkout's (possibly stale) current branch. `default_base` degrades to the
-    // current branch on a remote-less repo.
-    let base = match req.base.clone() {
-        Some(b) => b,
-        None => git::default_base(&repo_root).await?,
-    };
-    tracing::debug!(base = %base, "resolved base branch");
-
-    let (branch_name, work_dir) = if let Some(existing_branch) = existing {
-        tracing::info!(branch = %existing_branch, "reusing existing branch for session");
-        if !git::branch_exists(&repo_root, existing_branch).await {
-            return Err(AppError::bad_request(format!(
-                "branch '{existing_branch}' does not exist in this repo"
-            )));
-        }
-        // Reject if a tracked branch already has a live session.
-        if let Some(existing_b) =
-            branch_mod::find_by_repo_branch(&st.db, &repo_root_str, existing_branch).await?
-        {
-            if session_mod::active_for_branch(&st.db, &existing_b.id)
-                .await?
-                .is_some()
-            {
-                return Err(AppError::conflict(format!(
-                    "branch '{existing_branch}' already has an active session"
-                )));
-            }
-        }
-        let work_dir = match git::worktree_for_branch(&repo_root, existing_branch)
-            .await
-            .map_err(|e| AppError::bad_request(e.to_string()))?
-        {
-            Some(p) => {
-                tracing::debug!(branch = %existing_branch, work_dir = %p.display(), "found existing worktree for branch");
-                p
-            }
-            None => {
-                let slug = branch_mod::slugify(existing_branch);
-                let dir = repo_root.join(".worktrees").join(&slug);
-                tokio::fs::create_dir_all(repo_root.join(".worktrees")).await?;
-                git::ensure_excluded(&repo_root, ".worktrees/").await.ok();
-                tracing::info!(branch = %existing_branch, work_dir = %dir.display(), "provisioning worktree for existing branch");
-                git::worktree_add_existing(&repo_root, &dir, existing_branch)
-                    .await
-                    .map_err(|e| AppError::bad_request(e.to_string()))?;
-                dir
-            }
-        };
-        (existing_branch.to_string(), work_dir)
-    } else {
-        // Create `weaver/<slug>` with a unique suffix.
-        let explicit = req.name.as_deref().map(str::trim).filter(|n| !n.is_empty());
-        let base_slug = branch_mod::slugify(explicit.unwrap_or(title.as_str()));
-        tracing::debug!(base_slug = %base_slug, base = %base, "creating new branch for session");
-        let mut slug = base_slug.clone();
-        let mut suffix = 2;
-        loop {
-            let branch_name = format!("weaver/{slug}");
-            let dir = repo_root.join(".worktrees").join(&slug);
-            if !git::branch_exists(&repo_root, &branch_name).await && !dir.exists() {
-                break;
-            }
-            if explicit.is_some() {
-                return Err(AppError::conflict(format!(
-                    "a session named '{slug}' already exists — choose a different name"
-                )));
-            }
-            slug = format!("{base_slug}-{suffix}");
-            suffix += 1;
-        }
-        let branch_name = format!("weaver/{slug}");
-        let work_dir = repo_root.join(".worktrees").join(&slug);
-        tokio::fs::create_dir_all(repo_root.join(".worktrees")).await?;
-        git::ensure_excluded(&repo_root, ".worktrees/").await.ok();
-        tracing::info!(branch = %branch_name, work_dir = %work_dir.display(), base = %base, "provisioning worktree for new branch");
-        git::worktree_add(&repo_root, &work_dir, &branch_name, &base)
-            .await
-            .map_err(|e| AppError::bad_request(e.to_string()))?;
-        (branch_name, work_dir)
-    };
-
-    // A replacement launch may reuse a worktree whose previous session left
-    // Scratch files behind. Validate and write the merged set while holding the
-    // same path-scoped permit as live upload/delete routes, before creating any
-    // branch row, tracking issue, claim, or session.
-    let _scratch_permit = st.launch_gate.acquire_scratch(&work_dir).await;
-    let scratch_names = write_prepared_initial_scratch(&work_dir, &prepared_scratch).await?;
-
-    // Get-or-create the branch row, then stamp its title/goal/description.
-    let branch = branch_mod::upsert(&st.db, &repo_root_str, &branch_name, &base).await?;
-    tracing::debug!(branch = %branch.id, branch_name = %branch_name, "upserted branch row");
-    // A plain relaunch of an existing branch resumes its human-owned identity;
-    // only new task input intentionally replaces it.
-    let preserve_existing_title = existing.is_some()
-        && !title_was_explicit
-        && !issue_owned_title
-        && goal.is_empty()
-        && !branch.title.is_empty();
-    if preserve_existing_title {
-        title.clone_from(&branch.title);
-    } else {
-        branch_mod::set_title(&st.db, &branch.id, &title, title_provenance).await?;
-    }
-    if !goal.is_empty() {
-        branch_mod::set_goal(&st.db, &branch.id, &goal, "user").await?;
-    }
-    if !description.is_empty() {
-        branch_mod::set_description(&st.db, &branch.id, &description).await?;
-    }
-    // Re-fetch so the view we return reflects the freshly-stamped fields.
-    let branch = branch_mod::get(&st.db, &branch.id)
-        .await?
-        .ok_or_else(|| AppError::new(StatusCode::INTERNAL_SERVER_ERROR, "branch vanished"))?;
-    tracing::debug!(branch = %branch.id, title = %title, "stamped branch title/goal/description");
-
-    // Resolve the launching parent once: it names the tracking issue's
-    // `source_branch` *and* the session's tree parent (`parent_branch_id`).
-    // Only attribute to a parent in *this* repo, and never to the branch itself
-    // — `resolve_key` searches globally, so a stray `$WEAVER_BRANCH` from a
-    // checkout elsewhere must not misattribute the link to an unrelated branch.
-    let parent = match req
-        .parent_branch
-        .as_deref()
-        .map(str::trim)
-        .filter(|s| !s.is_empty())
-    {
-        Some(key) => branch_mod::resolve_key(&st.db, key)
-            .await?
-            .filter(|b| b.repo_root == branch.repo_root && b.branch != branch.branch),
-        None => None,
-    };
-    let parent_branch_name = parent.as_ref().map(|b| b.branch.clone());
-    let parent_session_id = match &parent {
-        Some(parent) => session_mod::active_for_branch(&st.db, &parent.id)
-            .await?
-            .map(|session| session.id),
-        None => None,
-    };
-    let stamped_allowed_tools = serde_json::to_string(&resolved.runtime_permissions)
-        .map_err(|error| AppError::bad_request(error.to_string()))?;
-    let stamped_mcp_access = serde_json::to_string(&resolved.mcp_policy)
-        .map_err(|error| AppError::bad_request(error.to_string()))?;
-    let (creator_kind, creator_subject) = actor.creator_identity();
-    let launch_policy = session_mod::SessionLaunchPolicy {
-        profile: profile_name.clone(),
-        launch_mode: mode.clone(),
-        profile_revision: launch_profile.revision,
-        profile_lifetime: launch_profile.lifetime,
-        strict: launch_profile.strict,
-        env_clear: launch_profile.env_clear,
-        ambient_allowlist: launch_profile.ambient_allowlist.clone(),
-        idle_archive_secs: resolved.view.policy.idle_archive_secs,
-        turn_budget: resolved.view.policy.turn_budget.unwrap_or(0),
-        prelude: launch_profile.prelude.clone(),
-        restricted: launch_profile.restricted,
-        allowed_tools: stamped_allowed_tools.clone(),
-        mcp_access: stamped_mcp_access,
-        launch_snapshot,
-        creator_kind: creator_kind.to_string(),
-        creator_subject,
-        parent_session_id,
-        automation_run_id: actor.automation_run_id().map(str::to_string),
-    };
-
-    // Open this session's tracking issue before the launch prompt is written,
-    // so the agent can be told its issue number. When an agent delegated this
-    // work (`parent_branch`), the parent becomes the issue's `source_branch`.
-    tracing::debug!(branch = %branch.id, "opening tracking issue for session");
-    let tracking_issue = create_tracking_issue(
-        &st,
-        &branch,
-        parent_branch_name.as_deref(),
-        &title,
-        &goal,
-        &description,
-        github_repo.as_deref(),
-        github_issue,
-        claimed_issue_id,
-    )
-    .await?;
-    tracing::debug!(branch = %branch.id, tracking_issue = ?tracking_issue, "tracking issue resolved");
-
-    // Automation reserves its session id durably before provisioning. Reusing
-    // it here makes retries converge on one runtime identity instead of
-    // silently allocating a second session after an ambiguous response.
-    let session_id = actor
-        .reserved_session_id()
-        .map(str::to_string)
-        .unwrap_or_else(branch_mod::new_id);
-    let run_dir = db::run_dir(&session_id);
-    tokio::fs::create_dir_all(&run_dir).await?;
-    tracing::info!(session = %session_id, branch = %branch.id, run_dir = %run_dir.display(), "allocated session id and run dir");
-
-    // Drop any attached reference files into the worktree before the agent
-    // launches, then tell the agent they are there. The branch goal stays the
-    // clean text the user typed; the scratch and tracking notes ride on the
-    // launch prompt (goal.txt) only, so they reach the agent without cluttering
-    // the dashboard.
-    tracing::debug!(session = %session_id, scratch_files = scratch_names.len(), "wrote initial scratch files");
-    // Goal, scratch, and tracking context ride in as the positional prompt that
-    // seeds the session's first turn.
-    let goal_file = {
-        let scratch = scratch_note(&scratch_names);
-        let entrance = entrance_note(tracking_issue);
-        let launch_prompt = build_launch_prompt(
-            &goal,
-            &launch_profile.prelude,
-            &entrance,
-            scratch.as_deref(),
-        );
-        if launch_prompt.is_empty() {
-            None
-        } else {
-            let f = run_dir.join("goal.txt");
-            tokio::fs::write(&f, &launch_prompt).await?;
-            tracing::debug!(session = %session_id, "wrote goal file for launch prompt");
-            Some(f)
-        }
-    };
-
-    let term_session = format!("weaver-{session_id}");
-    tracing::debug!(session = %session_id, term_session = %term_session, "derived terminal session name");
-
-    // Attribute the agent's commits to the launching user (design §6.3, Level A):
-    // export their GitHub identity as the git author/committer. Inserted only if
-    // not already set by a preceding env layer, so an explicit repo/operator
-    // override still wins, and only for an interactive launch that carries a
-    // `created_by` principal (webhook/warm/adopt paths have none and keep the
-    // shared identity).
-    if let Some(username) = created_by.as_deref() {
-        match crate::auth::commit_identity(&st.db, username).await {
-            Ok(Some(id)) => {
-                for (k, v) in [
-                    ("GIT_AUTHOR_NAME", &id.name),
-                    ("GIT_AUTHOR_EMAIL", &id.email),
-                    ("GIT_COMMITTER_NAME", &id.name),
-                    ("GIT_COMMITTER_EMAIL", &id.email),
-                ] {
-                    if !extra_env.iter().any(|(ek, _)| ek == k) {
-                        extra_env.push((k.to_string(), v.clone()));
-                    }
-                }
-                tracing::debug!(%username, "attributed commits to launching user");
-            }
-            Ok(None) => {
-                tracing::debug!(%username, "no commit identity registered, using shared identity")
-            }
-            Err(e) => tracing::warn!(%username, "failed to resolve commit identity: {e}"),
-        }
-    }
-
-    // Per-repo setup: run the repo's committed `[setup]` script in the worktree
-    // before the agent starts — but ONLY for an allowlisted (registered) repo,
-    // because a setup script is arbitrary, privileged code (it runs with the
-    // shared container's credentials; design §6.4). A non-allowlisted repo's
-    // script is never executed (recorded as skipped); a failed run leaves the
-    // session in a visible error state instead of launching a half-provisioned
-    // worktree.
-    if let Some(script) = repo_setup_for_profile(&repo_cfg, launch_profile.restricted) {
-        tracing::debug!(branch = %branch.id, repo = %repo_root.display(), "repo declares a [setup] script");
-        if repo::is_allowlisted(&st.db, &repo_root)
-            .await
-            .unwrap_or(false)
-        {
-            let outcome =
-                run_repo_setup(&st, &branch.id, &work_dir, &run_dir, &script, &extra_env).await;
-            if !outcome.success {
-                tracing::warn!(branch = %branch.id, "repo setup failed, aborting launch before agent start");
-                // Surface the failure as a loud, visible session state rather than
-                // launching the agent into a half-provisioned worktree. The
-                // worktree is left intact for inspection; full output is in the
-                // run dir's setup.log.
-                let session = insert_session_row(
-                    &st,
-                    &NewSession {
-                        id: session_id.clone(),
-                        branch_id: branch.id.clone(),
-                        work_dir: work_dir.display().to_string(),
-                        term_session: term_session.clone(),
-                        agent_kind: agent.clone(),
-                        model: model.clone(),
-                        effort: effort.clone(),
-                        status: "error".to_string(),
-                        github_repo: github_repo.clone(),
-                        parent_branch_id: parent.as_ref().map(|b| b.id.clone()),
-                        managed_by: None,
-                        created_by: created_by.clone(),
-                        protocol: protocol.clone(),
-                        origin: origin.to_string(),
-                        class: class.clone(),
-                        tracking_issue_id: tracking_issue,
-                    },
-                    &launch_policy,
-                )
-                .await?;
-                tracing::info!(
-                    branch = %branch.id,
-                    session = %session.id,
-                    status = %session.status,
-                    agent = %session.agent_kind,
-                    "session created"
-                );
-                let note = outcome.summary();
-                tags::set(
-                    &st.db,
-                    &branch.id,
-                    tags::ATTENTION_KEY,
-                    "blocked",
-                    &note,
-                    "loom",
-                )
-                .await
-                .ok();
-                events::record_tag(
-                    &st.db,
-                    &st.bus,
-                    &branch.id,
-                    tags::ATTENTION_KEY,
-                    "blocked",
-                    &note,
-                    "loom",
-                )
-                .await
-                .ok();
-                events::record(
-                    &st.db,
-                    &st.bus,
-                    &branch.id,
-                    "status",
-                    json!({ "status": "error", "reason": "repo setup failed" }),
-                )
-                .await
-                .ok();
-                return session_view(&st.db, &session, &branch).await;
-            }
-        } else {
-            tracing::info!(repo = %repo_root.display(),
-                "skipping .weaver/config.toml [setup]: repo is not allowlisted");
-            events::record(
-                &st.db,
-                &st.bus,
-                &branch.id,
-                "setup",
-                json!({ "phase": "skipped", "reason": "repo not allowlisted" }),
-            )
-            .await
-            .ok();
-        }
-    }
-
-    // Live the moment the agent spawns — there is no `launching` state.
-    let status = agent::initial_status(&st.db, &runtime).await;
-    let new_session = NewSession {
-        id: session_id.clone(),
-        branch_id: branch.id.clone(),
-        work_dir: work_dir.display().to_string(),
-        term_session: term_session.clone(),
-        agent_kind: agent.clone(),
-        model: model.clone(),
-        effort: effort.clone(),
-        status: status.to_string(),
-        github_repo: github_repo.clone(),
-        parent_branch_id: parent.as_ref().map(|b| b.id.clone()),
-        managed_by: None,
-        created_by: created_by.clone(),
-        protocol: protocol.clone(),
-        origin: origin.to_string(),
-        class: class.clone(),
-        tracking_issue_id: tracking_issue,
-    };
-    crate::auth::revoke_session_tokens(&st.db, &session_id).await?;
-    let session_token =
-        crate::auth::create_session_token(&st.db, created_by.as_deref(), &session_id, &branch.id)
-            .await?;
-    set_env(&mut extra_env, "LOOM_TOKEN", session_token);
-    let session = if protocol == "acp" {
-        // The ACP path inserts the row *first* — `acp::start` binds a relay to it
-        // and reads it back — then brings up the headless adapter over the relay.
-        tracing::info!(
-            session = %session_id, branch = %branch.id, runtime = %runtime,
-            work_dir = %work_dir.display(), mode = %mode, "launching acp session"
-        );
-        let session = insert_session_row(&st, &new_session, &launch_policy).await?;
-        // A custom ACP agent supplies the exact adapter command accepted by
-        // canonical resolution; registry edits after preview cannot replace it.
-        let launch = agent::build_acp_launch(
-            &st.db,
-            &agent::AcpLaunchSpec {
-                session_id: &session.id,
-                branch_id: &branch.id,
-                runtime: &runtime,
-                work_dir: &work_dir,
-                server_addr: &st.addr,
-                model: &model,
-                effort: &effort,
-                goal_file: goal_file.as_deref(),
-                primer_file: None,
-                extra_env: &extra_env,
-                env_clear: launch_profile.env_clear,
-                mode: &mode,
-                prelude: &launch_profile.prelude,
-                restricted: launch_profile.restricted,
-                allowed_tools: &stamped_allowed_tools,
-                mcp_access: &launch_policy.mcp_access,
-                custom: custom_agent.as_ref(),
-            },
-            agent::AcpOpen::Fresh,
-        )
-        .await
-        .map_err(|e| AppError::new(StatusCode::INTERNAL_SERVER_ERROR, e.to_string()))?;
-        if let Err(e) = crate::acp::start(&st, &session.id, launch).await {
-            crate::auth::revoke_session_tokens(&st.db, &session.id)
-                .await
-                .ok();
-            // Keep the durable row/worktree visible and recoverable, but retain
-            // non-2xx create semantics so CLI/webhook callers never announce a
-            // failed agent as successfully launched. The browser uses the
-            // returned session id to navigate to its handoff controls.
-            let _ = session_mod::set_status(&st.db, &session.id, "error").await;
-            let note =
-                format!("Agent failed to start: {e}. Hand off this session to another provider.");
-            tags::set(
-                &st.db,
-                &branch.id,
-                tags::ATTENTION_KEY,
-                "blocked",
-                &note,
-                "loom",
-            )
-            .await
-            .ok();
-            events::record_tag(
-                &st.db,
-                &st.bus,
-                &branch.id,
-                tags::ATTENTION_KEY,
-                "blocked",
-                &note,
-                "loom",
-            )
-            .await
-            .ok();
-            events::record(
-                &st.db,
-                &st.bus,
-                &branch.id,
-                "status",
-                json!({ "status": "error", "reason": "acp launch failed", "error": e.to_string() }),
-            )
-            .await
-            .ok();
-            return Err(
-                AppError::new(StatusCode::BAD_GATEWAY, format!("acp launch failed: {e}"))
-                    .with_fields(json!({ "session_id": session.id })),
-            );
-        }
-        tracing::info!(session = %session.id, branch = %branch.id, "acp session launched");
-        session
-    } else {
-        tracing::info!(
-            session = %session_id,
-            branch = %branch.id,
-            runtime = %runtime,
-            work_dir = %work_dir.display(),
-            env_vars = extra_env.len(),
-            "launching agent terminal"
-        );
-        // Make the session-bound token resolvable before the child starts. A
-        // terminal agent may call back into loom as soon as its shell execs;
-        // inserting after `agent::launch` left a race where that first request
-        // saw a correctly minted token as unauthorized.
-        let session = insert_session_row(&st, &new_session, &launch_policy).await?;
-        if let Err(e) = agent::launch(
-            &st.db,
-            &agent::LaunchSpec {
-                branch_id: &branch.id,
-                runtime: &runtime,
-                work_dir: &work_dir,
-                term_session: &term_session,
-                goal_file: goal_file.as_deref(),
-                primer_file: None,
-                prelude: &launch_profile.prelude,
-                server_addr: &st.addr,
-                model: &model,
-                effort: &effort,
-                extra_env: &extra_env,
-                env_clear: launch_profile.env_clear,
-                custom: custom_agent.as_ref(),
-            },
-            agent::LaunchMode::Fresh,
-        )
-        .await
-        {
-            crate::auth::revoke_session_tokens(&st.db, &session_id)
-                .await
-                .ok();
-            let _ = session_mod::set_status(&st.db, &session_id, "error").await;
-            return Err(AppError::new(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                e.to_string(),
-            ));
-        }
-        tracing::info!(session = %session_id, branch = %branch.id, "agent terminal launched");
-        session
-    };
-    tracing::debug!(session = %session.id, status = %status, "inserted session row");
-    // The next launch may begin once this agent is live. The remaining writes
-    // are bookkeeping and do not touch repository or worktree state.
-    drop(launch_permit);
-
-    if let Err(e) = repo::record_use(&st.db, &branch.repo_root).await {
-        tracing::warn!(branch = %branch.id, error = %e, "failed to record recent repo");
-    }
-    events::record(
-        &st.db,
-        &st.bus,
-        &branch.id,
-        "status",
-        json!({ "status": status, "reason": "session created" }),
-    )
-    .await
-    .ok();
-
-    tracing::info!(
-        branch = %branch.id,
-        session = %session.id,
-        status = %session.status,
-        agent = %session.agent_kind,
-        "session created"
-    );
-
-    crate::metadata_assist::spawn_title_generation(
-        st.db.clone(),
-        st.bus.clone(),
-        session.clone(),
-        branch.clone(),
-        false,
-    )
-    .await
-    .ok();
-    session_view(&st.db, &session, &branch).await
-}
-
-/// Session-specific operating context appended after the goal. Keep this
-/// compact: the goal is the outcome, the primer owns durable workflow rules,
-/// and this note only supplies the tracking contract that neither can know
-/// ahead of time. `weaver summary` is a recovery path, not a mandatory first
-/// tool call that would inject the goal a second time.
-fn entrance_note(tracking_issue: Option<i64>) -> String {
-    let mut note = "You are working in a Weaver session. Use `weaver summary` \
-                    to recover context if needed and `weaver readme` for the \
-                    complete workflow guide."
-        .to_string();
-    if let Some(id) = tracking_issue {
-        note.push_str(&format!(
-            " This session is tracked as weaver issue #{id}: keep `weaver \
-             status <level> \"<message>\"` honest as you work, and run `weaver \
-             issue close {id}` once the task is complete (e.g. the PR is open) \
-             so whoever launched you knows you are done."
-        ));
-    }
-    note
-}
-
-/// Construct the positional first prompt from the stamped prelude policy.
-/// The user's goal is always the opening user message: making an agent fetch it
-/// through `weaver summary` on turn one adds latency and duplicates the goal in
-/// context. `none` deliberately omits all Weaver orientation.
-fn build_launch_prompt(goal: &str, prelude: &str, entrance: &str, scratch: Option<&str>) -> String {
-    let mut parts = Vec::new();
-    if !goal.is_empty() {
-        parts.push(goal);
-        if prelude == "weaver" {
-            parts.push(entrance);
-        }
-    }
-    if let Some(scratch) = scratch {
-        parts.push(scratch);
-    }
-    parts.join("\n\n")
-}
-
-/// Open (or adopt) the tracking issue for a freshly-launched session: the one
-/// issue, claimed by the new branch, that represents its task. Whoever launched
-/// the session follows progress through it.
-///
-/// `--claim <id>` and `--issue <n>` (GitHub) reuse the issue they already
-/// imply, so a launch never opens a duplicate; a plain launch opens a fresh one
-/// from the task. An empty worktree with no task at all is untracked (`None`).
-/// `source_branch` records provenance — the parent branch when an agent
-/// delegated this work, else the new branch itself.
-#[allow(clippy::too_many_arguments)]
-async fn create_tracking_issue(
-    st: &AppState,
-    branch: &Branch,
-    parent_branch: Option<&str>,
-    title: &str,
-    goal: &str,
-    description: &str,
-    github_repo: Option<&str>,
-    github_issue: Option<i64>,
-    claim_issue: Option<i64>,
-) -> ApiResult<Option<i64>> {
-    let source = parent_branch.unwrap_or(&branch.branch).to_string();
-    tracing::debug!(branch = %branch.id, source = %source, "resolving tracking issue for session");
-
-    // Claiming an existing weaver issue: that issue *is* the tracker, so the
-    // claim must actually land — otherwise we'd hand back a tracking id for an
-    // issue this branch never claimed. Propagate failures rather than swallow.
-    if let Some(id) = claim_issue {
-        weaver_core::issue::set_claim(&st.db, id, Some(&branch.branch)).await?;
-        events::record(
-            &st.db,
-            &st.bus,
-            &branch.id,
-            "issue_claimed",
-            json!({ "id": id }),
-        )
-        .await
-        .ok();
-        return Ok(Some(id));
-    }
-
-    // A GitHub-seeded launch tracks the imported issue row.
-    if let Some(number) = github_issue {
-        let issue = weaver_core::issue::add(
-            &st.db,
-            &weaver_core::issue::NewIssue {
-                repo_root: branch.repo_root.clone(),
-                github_repo: github_repo.map(str::to_string),
-                source_branch: Some(source),
-                claimed_branch: Some(branch.branch.clone()),
-                title: title.to_string(),
-                body: description.to_string(),
-                github_issue: Some(number),
-            },
-        )
-        .await?;
-        events::record(
-            &st.db,
-            &st.bus,
-            &branch.id,
-            "issue_added",
-            json!({ "id": issue.id, "title": issue.title }),
-        )
-        .await
-        .ok();
-        return Ok(Some(issue.id));
-    }
-
-    // No task to track (e.g. an empty `--agent shell` worktree).
-    if goal.trim().is_empty() {
-        return Ok(None);
-    }
-
-    let body = if description.trim().is_empty() {
-        goal
-    } else {
-        description
-    };
-    let issue = weaver_core::issue::add(
-        &st.db,
-        &weaver_core::issue::NewIssue {
-            repo_root: branch.repo_root.clone(),
-            source_branch: Some(source),
-            claimed_branch: Some(branch.branch.clone()),
-            title: title.to_string(),
-            body: body.to_string(),
-            ..Default::default()
-        },
-    )
-    .await?;
-    events::record(
-        &st.db,
-        &st.bus,
-        &branch.id,
-        "issue_added",
-        json!({ "id": issue.id, "title": issue.title }),
-    )
-    .await
-    .ok();
-    Ok(Some(issue.id))
 }
 
 /// Set (upsert) a tag on a session's branch: validate `value` against the key's
@@ -2616,8 +1360,9 @@ pub(crate) async fn create_warm_session(
     // deliberately requires a live bound session, so an eager agent cannot hit
     // a transient authentication failure during startup.
     let status = agent::initial_status(&st.db, &agent).await;
-    let session = insert_session_row(
-        st,
+    let session = crate::session_layout::insert_session(
+        &st.db,
+        &st.bus,
         &NewSession {
             id: session_id.clone(),
             branch_id: branch.id.clone(),
@@ -4218,447 +2963,4 @@ pub(super) async fn set_mode(
     .await
     .ok();
     Ok(Json(json!({ "mode_id": req.mode_id })))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn legacy_create_preserves_explicit_agent_default_selectors() {
-        let req = CreateReq {
-            profile: Some(" template ".to_string()),
-            agent: Some(" ".to_string()),
-            model: Some(" ".to_string()),
-            effort: Some(String::new()),
-            protocol: Some(" ".to_string()),
-            mode: Some(" ".to_string()),
-            class: Some(" ".to_string()),
-            ..Default::default()
-        };
-        let selection = legacy_launch_selection(&req);
-        assert_eq!(selection.profile, "template");
-        assert_eq!(selection.overrides.agent, None);
-        assert_eq!(selection.overrides.model.as_deref(), Some(""));
-        assert_eq!(selection.overrides.effort.as_deref(), Some(""));
-        assert_eq!(selection.overrides.protocol.as_deref(), Some(""));
-        assert_eq!(selection.overrides.mode, None);
-        assert_eq!(selection.overrides.class, None);
-    }
-
-    async fn seed_user(db: &Db, username: &str) {
-        sqlx::query("INSERT INTO users (username) VALUES (?)")
-            .bind(username)
-            .execute(db)
-            .await
-            .unwrap();
-    }
-
-    struct EnvVarGuard {
-        name: &'static str,
-        value: Option<std::ffi::OsString>,
-    }
-
-    impl EnvVarGuard {
-        fn unset(name: &'static str) -> Self {
-            let value = std::env::var_os(name);
-            std::env::remove_var(name);
-            Self { name, value }
-        }
-
-        fn set(name: &'static str, new_value: &str) -> Self {
-            let value = std::env::var_os(name);
-            std::env::set_var(name, new_value);
-            Self { name, value }
-        }
-    }
-
-    impl Drop for EnvVarGuard {
-        fn drop(&mut self) {
-            match &self.value {
-                Some(value) => std::env::set_var(self.name, value),
-                None => std::env::remove_var(self.name),
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn user_github_token_injected_as_gh_token() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-        crate::user_token::set(&db, "alice", "ghp_alice")
-            .await
-            .unwrap();
-
-        let mut env = vec![("FOO".to_string(), "bar".to_string())];
-        apply_user_github_token(&db, &mut env, Some("alice")).await;
-        assert!(
-            env.iter().any(|(k, v)| k == "GH_TOKEN" && v == "ghp_alice"),
-            "the launching user's token is exported as GH_TOKEN"
-        );
-    }
-
-    #[tokio::test]
-    async fn restricted_environment_uses_only_profile_values() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        let dir = tempfile::tempdir().unwrap();
-        let repo = dir.path();
-        crate::profile::env_set(&db, "github_comment", "GH_TOKEN", "profile-token")
-            .await
-            .unwrap();
-        crate::repo_env::set(&db, &repo.display().to_string(), "REPO_SECRET", "leak")
-            .await
-            .unwrap();
-        let mut cfg = weaver_core::repo_config::RepoConfig::default();
-        cfg.env
-            .insert("COMMITTED_SECRET".to_string(), "leak".to_string());
-
-        let env = launch_env_for_profile(&db, repo, &cfg, "github_comment", true, true).await;
-
-        assert_eq!(
-            env,
-            vec![("GH_TOKEN".to_string(), "profile-token".to_string())]
-        );
-    }
-
-    #[tokio::test]
-    async fn stamped_strict_environment_keeps_profile_first_precedence() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        let dir = tempfile::tempdir().unwrap();
-        let repo = dir.path();
-        crate::profile::env_set(&db, "default", "SHARED_TOKEN", "profile")
-            .await
-            .unwrap();
-        crate::repo_env::set(
-            &db,
-            &repo.display().to_string(),
-            "SHARED_TOKEN",
-            "repository",
-        )
-        .await
-        .unwrap();
-        let mut cfg = weaver_core::repo_config::RepoConfig::default();
-        cfg.env
-            .insert("SHARED_TOKEN".to_string(), "committed".to_string());
-
-        let env = launch_env_for_profile(&db, repo, &cfg, "default", true, false).await;
-
-        assert_eq!(
-            env.iter()
-                .find(|(name, _)| name == "SHARED_TOKEN")
-                .map(|(_, value)| value.as_str()),
-            Some("profile")
-        );
-    }
-
-    #[tokio::test]
-    async fn ordinary_environment_keeps_cargo_target_worktree_local() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        let dir = tempfile::tempdir().unwrap();
-        let cfg = weaver_core::repo_config::RepoConfig::default();
-
-        let env = launch_env_for_profile(&db, dir.path(), &cfg, "default", false, false).await;
-
-        assert!(!env.iter().any(|(name, _)| name == "CARGO_TARGET_DIR"));
-    }
-
-    #[test]
-    fn restricted_profile_ignores_repository_setup() {
-        let mut cfg = weaver_core::repo_config::RepoConfig::default();
-        cfg.setup.script = Some("touch should-not-run".to_string());
-        assert_eq!(
-            repo_setup_for_profile(&cfg, false).as_deref(),
-            Some("touch should-not-run")
-        );
-        assert!(repo_setup_for_profile(&cfg, true).is_none());
-    }
-
-    #[tokio::test]
-    async fn user_token_overrides_ambient_gh_token_layer() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-        crate::user_token::set(&db, "alice", "ghp_alice")
-            .await
-            .unwrap();
-
-        // A lower env layer (the ambient Settings → Environment value, repo_env, …)
-        // already set GH_TOKEN: the user's own token overrides it *in place* — so
-        // their push/comment act as them — with no duplicate entry appended.
-        let mut env = vec![("GH_TOKEN".to_string(), "ambient-token".to_string())];
-        apply_user_github_token(&db, &mut env, Some("alice")).await;
-        let gh: Vec<&(String, String)> = env.iter().filter(|(k, _)| k == "GH_TOKEN").collect();
-        assert_eq!(gh.len(), 1, "no duplicate GH_TOKEN is appended");
-        assert_eq!(
-            gh[0].1, "ghp_alice",
-            "the user's own token wins over the ambient layer"
-        );
-    }
-
-    #[tokio::test]
-    async fn ambient_gh_token_is_the_fallback_without_a_user_token() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "bob").await; // bob has no stored token
-
-        // With no user token, whatever a lower layer set stands as the fallback.
-        let mut env = vec![("GH_TOKEN".to_string(), "ambient-token".to_string())];
-        apply_user_github_token(&db, &mut env, Some("bob")).await;
-        let gh: Vec<&(String, String)> = env.iter().filter(|(k, _)| k == "GH_TOKEN").collect();
-        assert_eq!(gh.len(), 1, "the ambient layer is left untouched");
-        assert_eq!(
-            gh[0].1, "ambient-token",
-            "with no user token, the ambient value is the fallback"
-        );
-    }
-
-    #[tokio::test]
-    async fn gh_token_untouched_without_token_or_principal() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-
-        // A user with no token set → nothing injected.
-        let mut env = vec![("FOO".to_string(), "bar".to_string())];
-        apply_user_github_token(&db, &mut env, Some("alice")).await;
-        assert!(!env.iter().any(|(k, _)| k == "GH_TOKEN"));
-
-        // A launch with no `created_by` (webhook/warm) → nothing injected, even
-        // though a token now exists.
-        crate::user_token::set(&db, "alice", "ghp_alice")
-            .await
-            .unwrap();
-        let mut env2 = vec![("FOO".to_string(), "bar".to_string())];
-        apply_user_github_token(&db, &mut env2, None).await;
-        assert!(!env2.iter().any(|(k, _)| k == "GH_TOKEN"));
-    }
-
-    #[serial_test::serial]
-    #[tokio::test]
-    async fn real_agent_requires_github_token_for_known_user() {
-        let _env = EnvVarGuard::unset("GH_TOKEN");
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-
-        let err = ensure_github_token_available(
-            &db,
-            &[("FOO".to_string(), "bar".to_string())],
-            Some("alice"),
-            "codex",
-            None,
-        )
-        .await
-        .unwrap_err();
-        assert_eq!(err.status(), StatusCode::PRECONDITION_REQUIRED);
-        assert_eq!(err.message(), crate::runtime::MISSING_GITHUB_TOKEN_MESSAGE);
-    }
-
-    #[serial_test::serial]
-    #[tokio::test]
-    async fn real_agent_accepts_user_or_default_github_token() {
-        let _env = EnvVarGuard::unset("GH_TOKEN");
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-
-        crate::user_token::set(&db, "alice", "ghp_alice")
-            .await
-            .unwrap();
-        ensure_github_token_available(&db, &[], Some("alice"), "claude", None)
-            .await
-            .unwrap();
-
-        crate::user_token::remove(&db, "alice").await.unwrap();
-        ensure_github_token_available(
-            &db,
-            &[("GH_TOKEN".to_string(), "ghp_shared".to_string())],
-            Some("alice"),
-            "codex",
-            None,
-        )
-        .await
-        .unwrap();
-    }
-
-    #[serial_test::serial]
-    #[tokio::test]
-    async fn empty_configured_gh_token_does_not_fall_back_to_ambient() {
-        let _env = EnvVarGuard::set("GH_TOKEN", "ghp_ambient");
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-
-        let err = ensure_github_token_available(
-            &db,
-            &[("GH_TOKEN".to_string(), " ".to_string())],
-            Some("alice"),
-            "codex",
-            None,
-        )
-        .await
-        .unwrap_err();
-        assert_eq!(err.status(), StatusCode::PRECONDITION_REQUIRED);
-    }
-
-    #[tokio::test]
-    async fn custom_and_webhook_launches_do_not_require_user_github_token() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-
-        // A custom (non-builtin) agent is exempt — it may never touch GitHub, and
-        // the operator supplies any credentials it needs via env.
-        ensure_github_token_available(&db, &[], Some("alice"), "my-custom-agent", None)
-            .await
-            .unwrap();
-        // A webhook launch carries an attribution string, not an approved user.
-        ensure_github_token_available(&db, &[], Some("github-webhook (octo)"), "codex", None)
-            .await
-            .unwrap();
-    }
-
-    #[serial_test::serial]
-    #[tokio::test]
-    async fn restricted_builtin_accepts_configured_github_app() {
-        let _env = EnvVarGuard::unset("GH_TOKEN");
-        let db = crate::db::connect_in_memory().await.unwrap();
-        seed_user(&db, "alice").await;
-        weaver_core::config::apply(
-            &db,
-            &[
-                (
-                    crate::github_app::APP_ID_KEY.to_string(),
-                    Some("123456".to_string()),
-                ),
-                (
-                    crate::github_app::APP_PRIVATE_KEY_KEY.to_string(),
-                    Some("configured-for-preflight".to_string()),
-                ),
-            ],
-        )
-        .await
-        .unwrap();
-        let app = crate::github_app::GithubApp::new(db.clone());
-
-        let err = ensure_github_token_available(&db, &[], Some("alice"), "claude", None)
-            .await
-            .unwrap_err();
-        assert_eq!(err.status(), StatusCode::PRECONDITION_REQUIRED);
-
-        ensure_github_token_available(&db, &[], Some("alice"), "claude", Some(&app))
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn create_tracking_issue_sources_parent_and_reuses_claims() {
-        let db = crate::db::connect_in_memory().await.unwrap();
-        let st = AppState {
-            db: db.clone(),
-            bus: crate::events::EventBus::new(),
-            addr: "127.0.0.1:0".to_string(),
-            ide: std::sync::Arc::new(crate::ide::IdeManager::new(crate::ide::ide_home())),
-            trigger: crate::github_trigger::GithubTrigger::production(db.clone()),
-            acp: crate::acp::AcpRegistry::new(),
-            launch_gate: crate::launch_gate::RepoLaunchGate::default(),
-        };
-        let child = branch_mod::upsert(&db, "/r", "weaver/child", "main")
-            .await
-            .unwrap();
-
-        // A delegated launch names the parent as the issue's source.
-        let id = create_tracking_issue(
-            &st,
-            &child,
-            Some("weaver/parent"),
-            "do it",
-            "do it in detail",
-            "",
-            None,
-            None,
-            None,
-        )
-        .await
-        .unwrap()
-        .expect("a fresh tracking issue");
-        let issue = weaver_core::issue::get(&db, id).await.unwrap().unwrap();
-        assert_eq!(issue.claimed_branch.as_deref(), Some("weaver/child"));
-        assert_eq!(
-            issue.source_branch.as_deref(),
-            Some("weaver/parent"),
-            "a delegated launch is sourced from the parent"
-        );
-
-        // A non-delegated launch is self-sourced (matches a hand-authored issue).
-        let id2 =
-            create_tracking_issue(&st, &child, None, "solo", "solo task", "", None, None, None)
-                .await
-                .unwrap()
-                .unwrap();
-        let issue2 = weaver_core::issue::get(&db, id2).await.unwrap().unwrap();
-        assert_eq!(issue2.source_branch.as_deref(), Some("weaver/child"));
-
-        // No task at all → nothing to track.
-        let none = create_tracking_issue(&st, &child, None, "", "", "", None, None, None)
-            .await
-            .unwrap();
-        assert!(none.is_none(), "an empty task opens no tracking issue");
-
-        // Claiming an existing issue reuses it rather than opening a duplicate.
-        let existing = weaver_core::issue::add(
-            &db,
-            &weaver_core::issue::NewIssue {
-                repo_root: "/r".to_string(),
-                title: "preexisting".to_string(),
-                ..Default::default()
-            },
-        )
-        .await
-        .unwrap();
-        let claimed = create_tracking_issue(
-            &st,
-            &child,
-            None,
-            "x",
-            "x",
-            "",
-            None,
-            None,
-            Some(existing.id),
-        )
-        .await
-        .unwrap();
-        assert_eq!(claimed, Some(existing.id), "a claim reuses the issue id");
-        let reclaimed = weaver_core::issue::get(&db, existing.id)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(
-            reclaimed.claimed_branch.as_deref(),
-            Some("weaver/child"),
-            "claiming stamps the new branch"
-        );
-    }
-
-    #[test]
-    fn entrance_note_keeps_catch_up_out_of_the_first_turn() {
-        let note = entrance_note(Some(42));
-        assert!(note.contains("weaver summary"));
-        assert!(note.contains("recover context"));
-        assert!(!note.contains("summary` first"));
-        assert!(!note.contains("prints the full goal"));
-        // It tells the agent exactly how to signal "done".
-        assert!(note.contains("weaver issue #42"));
-        assert!(note.contains("weaver issue close 42"));
-        assert!(note.contains("weaver status"));
-        // Untracked sessions get the orientation with no issue contract.
-        let untracked = entrance_note(None);
-        assert!(untracked.contains("weaver summary"));
-        assert!(!untracked.contains("issue"));
-    }
-
-    #[test]
-    fn restricted_prelude_delivers_the_caller_goal_without_weaver_orientation() {
-        let goal = "Rewrite only the issue body.\nBody hash: abc123";
-        let entrance = "weaver session metadata";
-        assert_eq!(build_launch_prompt(goal, "none", entrance, None), goal);
-        assert_eq!(
-            build_launch_prompt(goal, "weaver", entrance, None),
-            format!("{goal}\n\n{entrance}")
-        );
-    }
 }
